# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! 
    Directly from your PC, you can upload your photos, and they will be visible to everyone who is following you.
  version: "1"
  contact:
    email: fede.1942562@studenti.uniroma1.it
...
paths:
...
  /photos/{photo_id}/comments/{comment_id}:
    parameters:
    - $ref: "#/components/parameters/PhotoId"
    - name: comment_id
      description: The ID of the comment entry.
      in: path
      required: true
      schema:
        type: string
        example: "bcd456123"
        pattern: '^.*?$'
        minLength: 8
        maxLength: 16
    delete:
      tags:
      - comments
      summary: Remove a comment from a photo
      description: |-
        Remove the comment with the specified comment_id on the photo resource with photo_id, if created by the authenticated user.
      operationId: uncommentPhoto
      responses:
        "204":
          description: Comment deleted successfully.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: User can only delete his own comments
        "500": 
          $ref: "#/components/responses/InternalServerError"
      security:
        - bearerAuth: []
  /photos/{photo_id}/likes/{user_id}:
    parameters:
    - $ref: "#/components/parameters/PhotoId"
    - $ref: "#/components/parameters/UserId"
    put:
      tags:
      - likes
      summary: Set a like on a photo
      description: |- 
        Set a like on the photo resource with the specified photo_id on behalf of the user with the specified user_id.
        The user_id in the path must match the user ID of the authenticated user.
      operationId: likePhoto
      responses:
        "204":
          description: Photo liked successfully.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: Operation not allowed
        "404":
          description: Photo ID not found
        "500": 
          $ref: "#/components/responses/InternalServerError"
      security:
        - bearerAuth: []
    delete:
      tags:
      - likes
      summary: Remove a like on a photo
      description: |-
        Remove a like on the e with the specified photo_id on behalf of the user with the specified user_id (if the photo photo resourcwas previously liked by the user).
        The user_id in the path must match the user ID of the authenticated user.
      operationId: unlikePhoto
      responses:
        "204":
          description: Like removed successfully.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: User can only remove his own likes.
        "500": 
          $ref: "#/components/responses/InternalServerError"
      security:
        - bearerAuth: []
  /users:
    get:
      tags:
      - users
      summary: Search for a user by username
      description: |- 
        Search for a user by specifying their username.
        If no username is provided, then list all users.
      operationId: searchUser
      parameters:
      - name: username
        description: The username.
        in: query
        required: false
        schema:
          type: string
          example: Maria
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
      responses:
        "200":
          description: List of found users.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersInfoListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: Username not found
        "500": 
          $ref: "#/components/responses/InternalServerError"
      security:
        - bearerAuth: []
...
  /users/{user_id}:
    parameters:
    - $ref: "#/components/parameters/UserId"
    get:
      tags:
      - users
      summary: Get a user's profile
      description: |-
        Get information about the specified user's profile, specifically about the user's photos (in reverse chronological order), their followers and following numbers, and how many photos have been uploaded. 
      operationId: getUserProfile
      responses:
        "200":
          description: User's profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: User ID not found
        "500": 
          $ref: "#/components/responses/InternalServerError"
      security:
        - bearerAuth: []
    patch:
      tags:
      - users
      summary: Update the username
      description: |-
        Update the old username of the user resource with the specified user_id to the new username provided in the request body.
        The user_id in the path must match the user ID of the authenticated user.
      operationId: setMyUserName
      requestBody:
        content:
          application/json:
            schema:  
              $ref: "#/components/schemas/UserShortInfo"
      responses:
        "204":
          description: Username updated successfully.
        "400":  
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: User can only update his own username
        "409": # Conflict
          description: Username already taken
        "500": 
          $ref: "#/components/responses/InternalServerError"
      security:
        - bearerAuth: []
...
  /users/{user_id}/following/{target_user_id}:
    parameters:
    - $ref: "#/components/parameters/UserId"
    - $ref: "#/components/parameters/TargetUserId"
    delete:
      tags:
      - relationships
      summary: Allow a user to unfollow another user
      description: |-
        Allows the user with the specified user_id to unfollow the user with the specified target_user_id (if previously followed).
        The user_id in the path must match the user ID of the authenticated user.
      operationId: unfollowUser
      responses:
        "204":
          description: User unfollowed successfully.
        "400":  
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: Operation not allowed
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - bearerAuth: []
    put:
      tags:
      - relationships
      summary: Allow a user to follow another user
      description: |-
        Allows the user with the specified user_id to follow the user with the specified target_user_id.
        The user_id in the path must match the user ID of the authenticated user.
      operationId: followUser
      responses:
        "204":
          description: User followed successfully.
        "400":  
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: Operation not allowed
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - bearerAuth: []
...
  /users/{user_id}/banned/{target_user_id}:
    parameters:
    - $ref: "#/components/parameters/UserId"
    - $ref: "#/components/parameters/TargetUserId"
    delete:
      tags:
      - relationships
      summary: Allow a user to unban another user
      description: |-
        Allows the user with the specified user_id to unban the user with the specified target_user_id (if previously banned).
        The user_id in the path must match the user ID of the authenticated user.
      operationId: unbanUser
      responses:
        "204":
          description: User unbanned successfully.
        "400":  
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: Operation not allowed
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - bearerAuth: []
    put:
      tags:
      - relationships
      summary: Allow a user to ban another user
      description: |-
        Allows the user with the specified user_id to ban the user with the specified target_user_id.
        The user_id in the path must match the user ID of the authenticated user.
      operationId: banUser
      responses:
        "204":
          description: User banned successfully.
        "400":  
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: Operation not allowed
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - bearerAuth: []
components:
  responses:
    BadRequest:
      description: |-
        The request was not compliant with the documentation
        (eg. missing fields, etc).
    UnauthorizedError:
      description: |-
        Access token is missing or invalid
    InternalServerError:
      description: |-
        The server encountered an internal error.
        Further info in server logs.
  parameters:
    UserId:
      name: user_id
      in: path
      description: The ID of the user resource
      required: true
      schema:
        type: string
        example: "abc012345"
        pattern: '^.*?$'
        minLength: 8
        maxLength: 16
    PhotoId:
      name: photo_id
      in: path
      description: The ID of the photo resource
      required: true
      schema:
        type: string
        example: "def012345"
        pattern: '^.*?$'
        minLength: 8
        maxLength: 16
    TargetUserId:
      name: target_user_id
      in: path
      description: The ID of the target user resource.
      required: true
      schema:
        type: string
        example: "ghijkl67890"
        pattern: '^.*?$'
        minLength: 8
        maxLength: 16
  schemas:
...
    UserMetrics:
      description: Defines the public metrics of a user, namely the followers, following and uploaded photos number.
      type: object
      properties:
        followers:
          description: Number of a user's followers
          type: integer
          example: 28
          minimum: 0
        following:
          description: Number of users followed by a user
          type: integer
          example: 42
          minimum: 0
        photos:
          description: Number of photos uploaded by a user
          type: integer
          example: 16
          minimum: 0
    UserProfile:
      description: Defines a user profile, namely their id, username and public metrics.
      type: object
      properties:
        id:
          description: Unique user identifier
          type: string
          example: "abcdef012345"
        username:
          description: The username of the user
          type: string
          example: Maria
        metrics:
          $ref: "#/components/schemas/UserMetrics"
    UserShortInfo:
      description: Defines the short version of a user profile, just including their id and username.
      type: object
      properties:
        id:
          description: Unique user identifier
          type: string
          example: "abcdef012345"
          pattern: '^.*?$'
          minLength: 8
          maxLength: 16
          readOnly: true
        username:
          description: The username of the user
          type: string
          example: Maria
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
    UsersInfoListResponse:
      description: Defines a list of short-version user profiles.
      type: object
      properties:
        data:
          description: List of short-version user profiles.
          type: array
          items:
            $ref: "#/components/schemas/UserShortInfo"
          minItems: 0
          maxItems: 50
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer